refactor: moved domain models
