+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/models/attendance/AttendanceEntity.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.domain.models
+package jp.aoyama.mki.thermometer.domain.models.attendance
 
 import android.annotation.SuppressLint
 import java.text.SimpleDateFormat
