+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/models/device/BluetoothScanResult.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.domain.models
+package jp.aoyama.mki.thermometer.domain.models.device
 
 import java.util.*
