+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/models/temperature/BodyTemperatureEntity.kt
@@ -1,4 +1,4 @@
-package jp.aoyama.mki.thermometer.domain.models
+package jp.aoyama.mki.thermometer.domain.models.temperature
 
 import java.util.*
