+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/models/user/UserEntity.kt
@@ -1,4 +1,6 @@
-package jp.aoyama.mki.thermometer.domain.models
+package jp.aoyama.mki.thermometer.domain.models.user
+
+import jp.aoyama.mki.thermometer.domain.models.device.Device
 
 data class UserEntity(
     val id: String,
