+++ app/src/main/java/jp/aoyama/mki/thermometer/domain/repository/TemperatureRepository.kt
@@ -1,6 +1,6 @@
 package jp.aoyama.mki.thermometer.domain.repository
 
-import jp.aoyama.mki.thermometer.domain.models.BodyTemperatureEntity
+import jp.aoyama.mki.thermometer.domain.models.temperature.BodyTemperatureEntity
 
 interface TemperatureRepository {
     suspend fun findAll(): List<BodyTemperatureEntity>
