+++ app/src/main/java/jp/aoyama/mki/thermometer/view/bluetooth/list/BluetoothDiffUtil.kt
@@ -1,7 +1,7 @@
 package jp.aoyama.mki.thermometer.view.bluetooth.list
 
 import androidx.recyclerview.widget.DiffUtil
-import jp.aoyama.mki.thermometer.domain.models.BluetoothScanResult
+import jp.aoyama.mki.thermometer.domain.models.device.BluetoothScanResult
 
 class BluetoothDiffUtil : DiffUtil.ItemCallback<BluetoothScanResult>() {
     override fun areItemsTheSame(
