+++ app/src/main/java/jp/aoyama/mki/thermometer/view/temperature/list/BodyTemperatureDataDiffUtil.kt
@@ -1,7 +1,7 @@
 package jp.aoyama.mki.thermometer.view.temperature.list
 
 import androidx.recyclerview.widget.DiffUtil
-import jp.aoyama.mki.thermometer.domain.models.TemperatureData
+import jp.aoyama.mki.thermometer.domain.models.temperature.TemperatureData
 
 class BodyTemperatureDataDiffUtil : DiffUtil.ItemCallback<TemperatureData>() {
     override fun areItemsTheSame(oldItem: TemperatureData, newItem: TemperatureData): Boolean {
