+++ app/src/main/java/jp/aoyama/mki/thermometer/view/user/fragments/EditGradeFragment.kt
@@ -1,6 +1,5 @@
 package jp.aoyama.mki.thermometer.view.user.fragments
 
-import android.R
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -11,7 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import jp.aoyama.mki.thermometer.databinding.FragmentEditGradeBinding
-import jp.aoyama.mki.thermometer.domain.models.Grade
+import jp.aoyama.mki.thermometer.domain.models.user.Grade
 import jp.aoyama.mki.thermometer.view.user.viewmodels.CreateUserSharedViewModel
 
 class EditGradeFragment : Fragment() {
@@ -44,13 +43,13 @@ class EditGradeFragment : Fragment() {
             spinnerGrade.onItemSelectedListener = spinnerItemClickListener
             spinnerGrade.adapter = ArrayAdapter<String>(
                 requireContext(),
-                R.layout.simple_spinner_item,
+                android.R.layout.simple_spinner_item,
             ).apply {
                 val grades = Grade.values().map { it.gradeName }.toMutableList()
                 add("選択されていません")
                 addAll(grades)
 
-                setDropDownViewResource(R.layout.simple_spinner_dropdown_item)
+                setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item)
             }
         }
         return mBinding.root
