+++ app/src/main/java/jp/aoyama/mki/thermometer/view/user/viewmodels/CreateUserSharedViewModel.kt
@@ -2,10 +2,10 @@ package jp.aoyama.mki.thermometer.view.user.viewmodels
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
-import jp.aoyama.mki.thermometer.domain.models.Device
-import jp.aoyama.mki.thermometer.domain.models.Grade
-import jp.aoyama.mki.thermometer.domain.models.User
-import jp.aoyama.mki.thermometer.domain.models.UserEntity
+import jp.aoyama.mki.thermometer.domain.models.device.Device
+import jp.aoyama.mki.thermometer.domain.models.user.Grade
+import jp.aoyama.mki.thermometer.domain.models.user.User
+import jp.aoyama.mki.thermometer.domain.models.user.UserEntity
 import jp.aoyama.mki.thermometer.domain.service.UserService
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
